refactor: move commit, push, branch views into their own modules
