+++ src/main.rs
@@ -11,10 +11,16 @@ mod divider;
 pub(crate) use divider::{full_island_w, Divider, DIFF_GUTTER_W};
 mod ui;
 pub(crate) use ui::*;
+mod branch;
+mod commit;
 mod find;
 mod finder;
 mod history;
-pub(crate) use app::{CONTENT_MIN_QUERY, CONTENT_SEARCH_DEBOUNCE, FINDER_RESULT_CAP, SCROLL_CONTEXT_ROWS};
+mod push;
+pub(crate) use app::{
+    CONTENT_MIN_QUERY, CONTENT_SEARCH_DEBOUNCE, FINDER_RESULT_CAP, SCROLL_CONTEXT_ROWS,
+    STATUS_REFRESH_DEBOUNCE,
+};
 use kyde_diff as diff;
 mod editor;
 use kyde_config::keymap;
