+++ src/render.rs
@@ -709,481 +709,9 @@ impl Kyde {
         root.into_any_element()
     }
 
-    fn render_commit(
-        &mut self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        window: &mut Window,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let commit_n = self.files.len();
-        let push_n = self.push_files.len();
-        // Nothing to commit AND nothing to push → a single centered message.
-        if commit_n == 0 && push_n == 0 {
-            return div()
-                .flex()
-                .flex_1()
-                .h_full()
-                .items_center()
-                .justify_center()
-                .bg(t.main_bg)
-                .rounded(px(theme::ISLAND_RADIUS))
-                .font_family(ui)
-                .text_size(px(theme::get().ui_font_size + 1.0))
-                .text_color(t.line_number)
-                .child("You have nothing to commit or push")
-                .into_any_element();
-        }
 
-        // Only tabs with content are shown; fall back to the available one if the selected
-        // tab is the empty one (state is normalised after commit/push, this is a display guard).
-        let active = match self.git_tab {
-            GitTab::Commit if commit_n == 0 => GitTab::Push,
-            GitTab::Push if push_n == 0 => GitTab::Commit,
-            other => other,
-        };
-        // Tab bar (Commit / Push), then the active tab's left column + shared diff pane.
-        let tabs = self.render_git_tabs(active, cx);
-        // The Commit tab's files panel can be minimised to a thin strip (the `−` button in its
-        // header), handing the full width to the side-by-side diff.
-        let collapsed = active == GitTab::Commit && self.commit_collapsed;
-        // Auto-focus the commit-message box when the Commit view opens (deferred so the input
-        // element is in the tree first). Only when its tab is active and not collapsed — the
-        // box isn't rendered otherwise.
-        if self.focus_commit_msg && active == GitTab::Commit && !collapsed {
-            self.focus_commit_msg = false;
-            let handle = self.commit_editor.read(cx).focus_handle.clone();
-            window.focus(&handle);
-            window.defer(cx, move |window, _cx| window.focus(&handle));
-        }
-        let divider = div()
-            .id("commit-divider")
-            .w(px(theme::FRAME_GAP))
-            .h_full()
-            .flex_none()
-            .cursor_col_resize()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, e: &gpui::MouseDownEvent, window, cx| {
-                    this.start_divider_drag(Divider::Tree, e.position, window);
-                    cx.notify();
-                }),
-            );
-        // `min_w_0` is essential: without it the diff island (flex_1) sizes to its wide editor
-        // content instead of the remaining row width, and the last flex child (the right diff
-        // pane) collapses to zero — the side-by-side then shows only the base. History's diff
-        // works because its container already sets `min_w_0`.
-        let mut body = div().flex().flex_row().flex_1().min_h_0().min_w_0();
-        if collapsed {
-            // Thin strip with a `»` expand button where the files panel was (same affordance as
-            // the Browse tree's collapsed strip). No resize divider — nothing to resize.
-            body = body.child(
-                div()
-                    .flex_none()
-                    .h_full()
-                    .w(px(30.0))
-                    .py_1()
-                    .bg(t.panel_bg)
-                    .rounded(px(theme::ISLAND_RADIUS))
-                    .flex()
-                    .flex_col()
-                    .items_center()
-                    .child(
-                        div()
-                            .id("commit-expand")
-                            .flex()
-                            .items_center()
-                            .justify_center()
-                            .size(px(24.0))
-                            .rounded_md()
-                            .text_size(px(15.0))
-                            .text_color(t.line_number)
-                            .hover(|s| s.bg(t.bg_light).text_color(t.text))
-                            .cursor_pointer()
-                            .tooltip(|_w, cx| cx.new(|_| Tip("Expand files panel".into())).into())
-                            .child("»")
-                            .on_mouse_down(
-                                MouseButton::Left,
-                                cx.listener(|this, _e, _w, cx| {
-                                    this.commit_collapsed = false;
-                                    cx.notify();
-                                }),
-                            ),
-                    ),
-            );
-        } else {
-            let left = match active {
-                GitTab::Commit => self.render_commit_left(ui, cx),
-                GitTab::Push => self.render_push_left(ui, cx),
-            };
-            body = body.child(left).child(divider);
-        }
-        let body = body.child(self.render_diff(ui, fs, Some(window), cx));
 
-        div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .min_w_0()
-            .child(tabs)
-            .child(body)
-            .into_any_element()
-    }
 
-    /// Tab strip atop the git view: Commit (working changes) and Push (committed-but-unpushed),
-    /// each with a count badge when non-empty.
-    fn render_git_tabs(&self, active: GitTab, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let ui = theme::font::UI_FAMILY;
-        let mut row = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .flex_none()
-            .px_1()
-            .pb(px(theme::FRAME_GAP))
-            .font_family(ui)
-            .text_size(px(t.ui_font_size));
-        // A tab is shown only when it has files (reusable pill component, IntelliJ-style).
-        if !self.files.is_empty() {
-            row = row.child(
-                tab_pill(
-                    "git-tab-commit",
-                    "Commit",
-                    self.files.len(),
-                    active == GitTab::Commit,
-                )
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, _w, cx| this.set_git_tab(GitTab::Commit, cx)),
-                ),
-            );
-        }
-        if !self.push_files.is_empty() {
-            row = row.child(
-                tab_pill(
-                    "git-tab-push",
-                    "Push",
-                    self.push_files.len(),
-                    active == GitTab::Push,
-                )
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, _w, cx| this.set_git_tab(GitTab::Push, cx)),
-                ),
-            );
-        }
-        row.into_any_element()
-    }
-
-    /// Left column of the Commit tab: the changed-files tree (search + checkboxes) + the
-    /// commit-message bar.
-    fn render_commit_left(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let root_name = self
-            .repo_root
-            .as_ref()
-            .and_then(|p| p.file_name())
-            .map(|s| s.to_string_lossy().into_owned())
-            .unwrap_or_else(|| "/".to_string());
-        // Changed files as a folder tree (root + everything expanded by rebuild_commit_view).
-        let mut visible = vec![tree::Row {
-            path: PathBuf::new(),
-            is_dir: true,
-            depth: 0,
-        }];
-        if self.commit_expanded.contains(&PathBuf::new()) {
-            for mut r in self.commit_tree.visible(&self.commit_expanded) {
-                r.depth += 1;
-                visible.push(r);
-            }
-        }
-        // Filter the changed-files list by the search box: keep the root, files whose path
-        // matches, and folders that contain a matching file.
-        let query = self.commit_search.read(cx).text().trim().to_lowercase();
-        if !query.is_empty() {
-            let files = &self.files;
-            visible.retain(|r| {
-                r.path.as_os_str().is_empty()
-                    || (!r.is_dir && r.path.to_string_lossy().to_lowercase().contains(&query))
-                    || (r.is_dir
-                        && files.iter().any(|f| {
-                            f.path.starts_with(&r.path)
-                                && f.path.to_string_lossy().to_lowercase().contains(&query)
-                        }))
-            });
-        }
-        let rows: Vec<gpui::AnyElement> = visible
-            .into_iter()
-            .map(|r| {
-                let is_root = r.path.as_os_str().is_empty();
-                let checked = if r.is_dir {
-                    self.folder_all_checked(&r.path)
-                } else {
-                    self.commit_checked.contains(&r.path)
-                };
-                let file_idx = (!r.is_dir)
-                    .then(|| self.files.iter().position(|f| f.path == r.path))
-                    .flatten();
-                let selected = file_idx.is_some() && self.selected == file_idx;
-                let name_color = file_idx
-                    .and_then(|i| self.files.get(i))
-                    .map(|f| status_color(f.status))
-                    .unwrap_or(t.text);
-                let name: SharedString = if is_root {
-                    root_name.clone().into()
-                } else {
-                    r.path
-                        .file_name()
-                        .map(|n| n.to_string_lossy().into_owned())
-                        .unwrap_or_default()
-                        .into()
-                };
-                let expanded = self.commit_expanded.contains(&r.path);
-                let is_dir = r.is_dir;
-                let (p_act, p_check, p_ctx) = (r.path.clone(), r.path.clone(), r.path.clone());
-                self.tree_row(
-                    cx,
-                    &r.path,
-                    is_dir,
-                    expanded,
-                    r.depth,
-                    selected,
-                    name,
-                    name_color,
-                    Some(checked),
-                    move |this, _e, _w, cx| {
-                        if is_dir {
-                            this.toggle_commit_dir(p_act.clone(), cx);
-                        } else if let Some(i) = this.files.iter().position(|f| f.path == p_act) {
-                            this.select_with(i, Some(cx));
-                            cx.notify();
-                        }
-                    },
-                    move |this, cx| this.toggle_commit_check(p_check.clone(), is_dir, cx),
-                    move |this, pos, cx| {
-                        this.open_menu(pos, MenuTarget::CommitPath(p_ctx.clone(), is_dir), cx);
-                    },
-                )
-            })
-            .collect();
-        // File-list island (same island styling/width as the Browse tree): a fixed search
-        // header (filter box + divider) over the scrollable changed-files list.
-        // Header: filter box + a `−` button that minimises the panel to a thin strip (same
-        // affordance as the Browse tree), handing the full width to the side-by-side diff.
-        let collapse_btn = div()
-            .id("commit-minimize")
-            .flex_none()
-            .flex()
-            .items_center()
-            .justify_center()
-            .size(px(22.0))
-            .rounded_md()
-            .text_size(px(16.0))
-            .text_color(t.line_number)
-            .hover(|s| s.bg(t.bg_light).text_color(t.text))
-            .cursor_pointer()
-            .tooltip(|_w, cx| cx.new(|_| Tip("Collapse files panel".into())).into())
-            .child("−")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.commit_collapsed = true;
-                    cx.notify();
-                }),
-            );
-        let search_header = div()
-            .flex_none()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_1()
-            .px_2()
-            .py_1p5()
-            .text_size(px(theme::get().ui_font_size))
-            .child(div().flex_1().min_w_0().child(self.commit_search.clone()))
-            .child(collapse_btn);
-        let search_hr = div().flex_none().h(px(1.0)).mx_1().bg(t.divider);
-        let file_list = div()
-            .id("commit-tree")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .px_1()
-            .py_1()
-            .children(rows);
-        let list_island = div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .text_color(t.text)
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size + 1.0))
-            .child(search_header)
-            .child(search_hr)
-            .child(file_list);
-
-        // Left column: file list + commit message, the same width as the Browse tree.
-        div()
-            .flex()
-            .flex_col()
-            .gap(px(theme::FRAME_GAP))
-            .w(px(self.tree_width))
-            .flex_none()
-            .h_full()
-            .child(list_island)
-            .child(self.render_commit_bar(cx))
-            .into_any_element()
-    }
-
-    /// Left column of the Push tab: a flat list of the files a push would send (click → diff)
-    /// + a footer with the Push button. No checkboxes — a push is all-or-nothing.
-    fn render_push_left(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        let n = self.push_files.len();
-        let rows: Vec<gpui::AnyElement> = self
-            .push_files
-            .iter()
-            .enumerate()
-            .map(|(i, f)| {
-                let selected = self.push_selected == Some(i);
-                let name: SharedString = f
-                    .path
-                    .file_name()
-                    .map(|s| s.to_string_lossy().into_owned())
-                    .unwrap_or_else(|| f.path.to_string_lossy().into_owned())
-                    .into();
-                let dir = f
-                    .path
-                    .parent()
-                    .map(|p| p.to_string_lossy().into_owned())
-                    .filter(|s| !s.is_empty());
-                let name_color = status_color(f.status);
-                let path = f.path.clone();
-                div()
-                    .id(("push-file", i))
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .mx(px(6.0))
-                    .px_2()
-                    .py_1()
-                    .rounded_md()
-                    .cursor_pointer()
-                    .when(selected, |d| d.bg(t.selected_bg))
-                    .when(!selected, |d| d.hover(|s| s.bg(t.bg_mid)))
-                    .child(div().flex_none().child(badge_inner(file_badge(&path), 2.0)))
-                    .child(div().flex_none().text_color(name_color).child(name))
-                    .when_some(dir, |d, dir| {
-                        d.child(
-                            div()
-                                .flex_1()
-                                .min_w_0()
-                                .truncate()
-                                .text_color(t.line_number)
-                                .child(SharedString::from(dir)),
-                        )
-                    })
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(move |this, _e, _w, cx| this.select_push_file(i, cx)),
-                    )
-                    .into_any_element()
-            })
-            .collect();
-
-        let count_header = div()
-            .flex_none()
-            .px_3()
-            .py_1p5()
-            .text_color(t.line_number)
-            .child(SharedString::from(if n == 0 {
-                "Nothing to push".to_string()
-            } else {
-                format!("{n} file{} to push", if n == 1 { "" } else { "s" })
-            }));
-        let list = div()
-            .id("push-list")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .py_1()
-            .children(rows);
-        let list_island = div()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .min_h_0()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .text_color(t.text)
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size + 1.0))
-            .child(count_header)
-            .child(div().flex_none().h(px(1.0)).mx_1().bg(t.divider))
-            .child(list);
-
-        // Footer bar: Cancel (back to Browse) + Push, styled like the commit bar's buttons.
-        let cancel_btn = div()
-            .px_4()
-            .py_2()
-            .rounded_md()
-            .border_1()
-            .border_color(t.divider)
-            .text_color(t.secondary_text)
-            .cursor_pointer()
-            .child("Cancel")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.mode = Mode::Browse;
-                    cx.notify();
-                }),
-            );
-        let pushing = self.pushing;
-        let push_btn =
-            btn_primary_state("push", if pushing { "Pushing…" } else { "Push" }, pushing)
-                .py_2()
-                .font_weight(FontWeight::SEMIBOLD)
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, _w, cx| this.do_push(cx)),
-                );
-        let bar = div()
-            .flex()
-            .flex_row()
-            .justify_end()
-            .gap_2()
-            .flex_none()
-            .p_2()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .font_family(ui)
-            .child(cancel_btn)
-            .when(n > 0, |d| d.child(push_btn));
-
-        div()
-            .flex()
-            .flex_col()
-            .gap(px(theme::FRAME_GAP))
-            .w(px(self.tree_width))
-            .flex_none()
-            .h_full()
-            .child(list_island)
-            .child(bar)
-            .into_any_element()
-    }
 
 
     /// Side-by-side diff = two editors in one rounded island: left is the read-only base
@@ -1750,76 +1278,6 @@ impl Kyde {
         )
     }
 
-    fn render_commit_bar(&self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        // Cancel → back to the Browse (code) view.
-        let cancel_btn = div()
-            .px_4()
-            .py_2()
-            .rounded_md()
-            .border_1()
-            .border_color(t.divider)
-            .text_color(t.secondary_text)
-            .child("Cancel")
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| {
-                    this.mode = Mode::Browse;
-                    cx.notify();
-                }),
-            );
-        // Emphasized primary CTA: standard primary button + taller pad + semibold. While a
-        // commit is in flight it's dimmed + labelled "Committing…" (clicks are no-ops).
-        let committing = self.committing;
-        let commit_btn = btn_primary_state(
-            "commit",
-            if committing {
-                "Committing…"
-            } else {
-                "Commit"
-            },
-            committing,
-        )
-        .py_2()
-        .font_weight(FontWeight::SEMIBOLD)
-        .on_mouse_down(
-            MouseButton::Left,
-            cx.listener(|this, _e, _w, cx| {
-                this.commit_now(cx);
-                cx.notify();
-            }),
-        );
-
-        div()
-            .flex()
-            .flex_col()
-            .gap_2()
-            .h(px(150.0))
-            .flex_none()
-            .p_2()
-            .bg(t.panel_bg)
-            .rounded(px(theme::ISLAND_RADIUS))
-            .child(
-                div()
-                    .flex_1()
-                    .min_h_0()
-                    .p_2()
-                    .bg(t.bg_mid)
-                    .rounded_md()
-                    .child(self.commit_editor.clone()),
-            )
-            // Cancel + Commit on their own line, right-aligned.
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .justify_end()
-                    .gap_2()
-                    .child(cancel_btn)
-                    .child(commit_btn),
-            )
-            .into_any_element()
-    }
 
     /// Shared tree row used by Browse, Commit, and Rollback — so items look identical
     /// everywhere. `checkbox: None` = no checkbox; `Some(checked)` shows one. The three
@@ -3676,470 +3134,11 @@ impl Kyde {
         col.into_any_element()
     }
 
-    /// Bottom status bar — currently just the branch switcher, anchored bottom-right.
-    fn render_status_bar(&self, ui: &'static str, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let t = theme::get();
-        // All bottom-bar text is this muted grey; icons keep their own colours.
-        let bar_text = gpui::rgb(0x808289);
-        let label = self
-            .current_branch
-            .clone()
-            .unwrap_or_else(|| "(no branch)".into());
-        let chip = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .flex_none()
-            .gap_1p5()
-            .px_2()
-            .rounded_md()
-            .cursor_pointer()
-            .hover(|s| s.bg(t.bg_light))
-            .text_color(bar_text)
-            .child(
-                div().flex_none().child(
-                    svg()
-                        .path("icons/git-branch.svg")
-                        .size(px(15.0))
-                        .text_color(bar_text),
-                ),
-            )
-            .child(SharedString::from(label))
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, window, cx| this.toggle_branch_popup(window, cx)),
-            );
 
-        // Breadcrumb of the open file: <repo> › dir › … › <badge> file. flex_1 + min_w_0
-        // + overflow hidden lets it shrink and clip rather than push into the branch chip.
-        let mut crumbs = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_1()
-            .flex_1()
-            .min_w_0()
-            .overflow_hidden()
-            // Line the first crumb up with the rail button icons (rail margin + icon inset).
-            .pl(px(8.0));
-        // Breadcrumb follows the tree selection (folder or file), falling back to the open
-        // file. A folder selection shows a folder icon; a file shows its type badge.
-        let crumb = self.selected_path.as_ref().or(self.open_path.as_ref());
-        if let Some(rel) = crumb.filter(|p| !p.as_os_str().is_empty()) {
-            // Real filesystem check (not "is it the open file") — single-click selection
-            // can point at any file or folder, so the icon must follow the path itself.
-            let is_file = self
-                .repo_root
-                .as_ref()
-                .map(|root| root.join(rel).is_file())
-                .unwrap_or(false);
-            let repo_name = self
-                .repo_root
-                .as_ref()
-                .and_then(|p| p.file_name())
-                .map(|s| s.to_string_lossy().into_owned())
-                .unwrap_or_default();
-            let sep = || div().flex_none().text_color(bar_text).child("›");
-            let folder_icon = || {
-                div().flex_none().child(
-                    svg()
-                        .path("icons/folder.svg")
-                        .size(px(16.0))
-                        .text_color(t.line_number),
-                )
-            };
-            crumbs = crumbs.child(
-                div()
-                    .flex_none()
-                    .text_color(bar_text)
-                    .child(SharedString::from(repo_name)),
-            );
-            let comps: Vec<String> = rel
-                .components()
-                .map(|c| c.as_os_str().to_string_lossy().into_owned())
-                .collect();
-            let last = comps.len().saturating_sub(1);
-            for (i, name) in comps.into_iter().enumerate() {
-                crumbs = crumbs.child(sep());
-                // Icon: file-type badge for the open file's final segment; folder icon
-                // for every directory segment (including a selected folder).
-                let icon = if i == last && is_file {
-                    div()
-                        .flex_none()
-                        .child(badge_inner(file_badge(rel), 2.0))
-                        .into_any_element()
-                } else {
-                    folder_icon().into_any_element()
-                };
-                crumbs = crumbs.child(icon).child(
-                    div()
-                        .flex_none()
-                        .text_color(bar_text)
-                        .child(SharedString::from(name)),
-                );
-            }
-        }
 
-        // Push button: ↑ + "Push", with an ahead-of-upstream count badge. Tooltip
-        // carries the last push error (or a hint). Disabled while a push is running.
-        let pushing = self.pushing;
-        let ahead = self.ahead.unwrap_or(0);
-        let tip_text: SharedString = self
-            .push_msg
-            .clone()
-            .map(SharedString::from)
-            .unwrap_or_else(|| "Push to origin".into());
-        let push_btn = div()
-            .id("push-btn")
-            .flex()
-            .flex_row()
-            .items_center()
-            .flex_none()
-            .gap_1p5()
-            .px_2()
-            .rounded_md()
-            .cursor_pointer()
-            .hover(|s| s.bg(t.bg_light))
-            .text_color(if self.push_msg.is_some() {
-                t.status_deleted
-            } else {
-                bar_text
-            })
-            .tooltip(move |_w, cx| cx.new(|_| Tip(tip_text.clone())).into())
-            .child(div().flex_none().child(if pushing { "↻" } else { "↑" }))
-            .child(SharedString::from(if pushing {
-                "Pushing…"
-            } else {
-                "Push"
-            }))
-            .when(ahead > 0, |d| {
-                d.child(
-                    div()
-                        .flex_none()
-                        .px_1p5()
-                        .rounded_md()
-                        .bg(t.bg_light)
-                        .text_color(t.text)
-                        .child(SharedString::from(ahead.to_string())),
-                )
-            })
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.open_push_modal(cx)),
-            );
-
-        // Pull chip: ↓ + "Pull", with a behind-of-upstream count badge. Shown only when we
-        // know we're behind (or a pull's in flight); the branch popup always offers Pull.
-        let pulling = self.pulling;
-        let behind = self.behind.unwrap_or(0);
-        let pull_btn = div()
-            .id("pull-btn")
-            .flex()
-            .flex_row()
-            .items_center()
-            .flex_none()
-            .gap_1p5()
-            .px_2()
-            .rounded_md()
-            .cursor_pointer()
-            .hover(|s| s.bg(t.bg_light))
-            .text_color(bar_text)
-            .tooltip(|_w, cx| cx.new(|_| Tip("Pull from origin (rebase)".into())).into())
-            .child(div().flex_none().child(if pulling { "↻" } else { "↓" }))
-            .child(SharedString::from(if pulling {
-                "Pulling…"
-            } else {
-                "Pull"
-            }))
-            .when(behind > 0, |d| {
-                d.child(
-                    div()
-                        .flex_none()
-                        .px_1p5()
-                        .rounded_md()
-                        .bg(t.bg_light)
-                        .text_color(t.text)
-                        .child(SharedString::from(behind.to_string())),
-                )
-            })
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, _w, cx| this.do_pull(cx)),
-            );
-
-        div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .justify_between()
-            .gap_2()
-            .h(px(28.0))
-            .mb(px(6.0))
-            .px_2()
-            // Joins the surrounding chrome: same chrome colour, no separating border.
-            .bg(t.frame_bg)
-            .font_family(ui)
-            // Same size as the file-tree rows.
-            .text_size(px(theme::get().ui_font_size + 3.0))
-            .child(crumbs)
-            .child(
-                div()
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .flex_none()
-                    .gap_2()
-                    // Only show Pull when we know we're behind (or one's in flight).
-                    .when(self.behind.unwrap_or(0) > 0 || self.pulling, |d| {
-                        d.child(pull_btn)
-                    })
-                    // Only show Push when there's actually something to push (or one's in flight).
-                    .when(self.ahead.unwrap_or(0) > 0 || self.pushing, |d| {
-                        d.child(push_btn)
-                    })
-                    .child(chip),
-            )
-            .into_any_element()
-    }
-
-    /// Branch switcher popup: search box, New Branch, Recent, then All Branches.
-    /// Anchored bottom-right above the status bar; transparent backdrop closes it.
-    fn render_branch_popup(
-        &self,
-        ui: &'static str,
-        fs: gpui::Pixels,
-        cx: &mut Context<Self>,
-    ) -> gpui::AnyElement {
-        let t = theme::get();
-        let q = self.branch_query.read(cx).text().to_string();
-        let ql = q.trim().to_lowercase();
-        let current = self.current_branch.clone();
-        let matches = |b: &str| ql.is_empty() || b.to_lowercase().contains(&ql);
-
-        let recent: Vec<String> = self
-            .branch_list
-            .iter()
-            .filter(|b| current.as_deref() != Some(b.as_str()))
-            .filter(|b| matches(b))
-            .take(5)
-            .cloned()
-            .collect();
-        let mut all: Vec<String> = self
-            .branch_list
-            .iter()
-            .filter(|b| matches(b))
-            .cloned()
-            .collect();
-        all.sort_by_key(|b| b.to_lowercase());
-
-        let nb_label = if ql.is_empty() {
-            "+ New Branch".to_string()
-        } else {
-            format!("+ New Branch  “{}”", q.trim())
-        };
-        // Popup separators: the theme `divider` (#26272B) is invisible on the popup's
-        // `bg_mid` (#26282B), so use a faint white hairline that actually reads.
-        let sep = gpui::rgba(0xFFFFFF1A);
-        let new_row = div()
-            .mx_1()
-            .px_2()
-            .py_1()
-            .rounded_md()
-            .cursor_pointer()
-            .hover(|s| s.bg(t.selected_bg))
-            .text_color(t.primary)
-            .child(SharedString::from(nb_label))
-            // Opens the "Create New Branch" dialog (the search text prefills the name).
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, _e, _w, cx| this.open_new_branch(cx)),
-            );
-
-        // Pull (fetch + rebase). Always available here — it's the repo-ops hub, and a pull
-        // fetches first, so it works even when our last-known `behind` count is stale/0.
-        let behind = self.behind.unwrap_or(0);
-        let pull_label = if self.pulling {
-            "↓ Pulling…".to_string()
-        } else if behind > 0 {
-            format!("↓ Pull  ({behind})")
-        } else {
-            "↓ Pull".to_string()
-        };
-        let pull_row = div()
-            .mx_1()
-            .px_2()
-            .py_1()
-            .rounded_md()
-            .cursor_pointer()
-            .hover(|s| s.bg(t.selected_bg))
-            .text_color(t.text)
-            .child(SharedString::from(pull_label))
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(move |this, _e, _w, cx| this.do_pull(cx)),
-            );
-
-        let search = div()
-            .flex()
-            .flex_row()
-            .items_center()
-            .gap_2()
-            .px_3()
-            .py_2()
-            .border_b_1()
-            .border_color(sep)
-            .child(
-                svg()
-                    .path("icons/search.svg")
-                    .size(px(15.0))
-                    .flex_none()
-                    .text_color(t.line_number),
-            )
-            .child(div().flex_1().min_w_0().child(self.branch_query.clone()));
-
-        // Branch tree: Recent + Local sections as expandable roots; `/` → folders.
-        // While searching, force everything open so matches are visible.
-        let rows = branch_rows(&recent, &all, &self.branch_expanded, !ql.is_empty());
-        let tree_rows = self.branch_tree(rows, cx);
-        let list = div()
-            .id("branch-list")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .py_1()
-            .child(new_row)
-            .child(pull_row)
-            // hairline between the actions and the branch tree
-            .child(div().mx_2().my_1().h(px(1.0)).bg(sep))
-            .children(tree_rows);
-
-        let panel = div()
-            .absolute()
-            .right(px(8.0))
-            .bottom(px(28.0))
-            .w(px(340.0))
-            .max_h(px(460.0))
-            .flex()
-            .flex_col()
-            .bg(t.bg_mid)
-            .border_1()
-            .border_color(gpui::rgb(0x595D60))
-            .rounded_md()
-            .shadow_lg()
-            .occlude()
-            .font_family(ui)
-            .text_size(fs)
-            .child(search)
-            .child(list);
-
-        div()
-            .absolute()
-            .top_0()
-            .left_0()
-            .size_full()
-            .on_mouse_down(
-                MouseButton::Left,
-                cx.listener(|this, _e, window, cx| {
-                    this.branch_popup_open = false;
-                    window.focus(&this.focus_handle);
-                    cx.notify();
-                }),
-            )
-            .child(panel)
-            .into_any_element()
-    }
-
-    /// Render the branch tree (sections as roots, `/` segments as folders).
-    fn branch_tree(&self, rows: Vec<BranchRow>, cx: &mut Context<Self>) -> Vec<gpui::AnyElement> {
-        let t = theme::get();
-        let current = self.current_branch.clone();
-        rows.into_iter()
-            .map(|r| {
-                let indent = px(8.0 + r.depth as f32 * 14.0);
-                match r.node {
-                    BranchNode::Folder {
-                        key,
-                        expanded,
-                        section,
-                    } => {
-                        let k = key.clone();
-                        div()
-                            .flex()
-                            .flex_row()
-                            .items_center()
-                            .gap_1()
-                            .h(px(28.0))
-                            .pl(indent)
-                            .pr_2()
-                            .rounded_md()
-                            .cursor_pointer()
-                            .hover(|s| s.bg(t.bg_light))
-                            .text_color(t.text)
-                            .child(
-                                div()
-                                    .w(px(12.0))
-                                    .flex_none()
-                                    .text_color(t.line_number)
-                                    .child(if expanded { "▾" } else { "▸" }),
-                            )
-                            .when(!section, |d| {
-                                d.child(
-                                    svg()
-                                        .path("icons/folder.svg")
-                                        .size(px(14.0))
-                                        .text_color(t.line_number),
-                                )
-                            })
-                            .child(SharedString::from(r.label))
-                            .on_mouse_down(
-                                MouseButton::Left,
-                                cx.listener(move |this, _e, _w, cx| {
-                                    this.toggle_branch_node(k.clone(), cx)
-                                }),
-                            )
-                            .into_any_element()
-                    }
-                    BranchNode::Leaf { full } => {
-                        let is_current = current.as_deref() == Some(full.as_str());
-                        let nm = full.clone();
-                        div()
-                            .flex()
-                            .flex_row()
-                            .items_center()
-                            .gap_2()
-                            .h(px(28.0))
-                            .pl(indent)
-                            .pr_2()
-                            .rounded_md()
-                            .cursor_pointer()
-                            .hover(|s| s.bg(t.selected_bg))
-                            .text_color(t.text)
-                            .child(div().flex_none().text_color(t.line_number).child("⎇"))
-                            .child(
-                                div()
-                                    .flex_1()
-                                    .min_w_0()
-                                    .truncate()
-                                    .child(SharedString::from(r.label)),
-                            )
-                            .when(is_current, |d| {
-                                d.child(div().flex_none().text_color(t.line_number).child("✓"))
-                            })
-                            .on_mouse_down(
-                                MouseButton::Left,
-                                cx.listener(move |this, _e, window, cx| {
-                                    this.checkout_branch(nm.clone(), window, cx)
-                                }),
-                            )
-                            .into_any_element()
-                    }
-                }
-            })
-            .collect()
-    }
 
     /// Right-click menu, positioned at the cursor. Transparent backdrop closes it.
-    fn render_context_menu(&self, cx: &mut Context<Self>) -> gpui::AnyElement {
+    pub(crate) fn render_context_menu(&self, cx: &mut Context<Self>) -> gpui::AnyElement {
         let Some(menu) = &self.context_menu else {
             return div().into_any_element();
         };
@@ -4458,138 +3457,6 @@ impl Kyde {
             .into_any_element()
     }
 
-    /// Delete-confirmation modal: name the target, Cancel / Delete.
-    /// Push confirmation modal: branch + the commits that would be pushed, Cancel / Push.
-    /// Push confirmation window body (own native window; titlebar shows "Push").
-    pub(crate) fn render_push_body(&mut self, cx: &mut Context<Self>) -> gpui::AnyElement {
-        let ui = theme::font::UI_FAMILY;
-        let t = theme::get();
-        let n = self.push_files.len();
-        // Flat file list, styled like the commit/rollback rows (badge + name + status
-        // color), right-click → "View Diff". No checkboxes — a push is all-or-nothing.
-        let rows: Vec<gpui::AnyElement> = self
-            .push_files
-            .iter()
-            .enumerate()
-            .map(|(i, f)| {
-                let name: SharedString = f
-                    .path
-                    .file_name()
-                    .map(|s| s.to_string_lossy().into_owned())
-                    .unwrap_or_else(|| f.path.to_string_lossy().into_owned())
-                    .into();
-                let dir = f
-                    .path
-                    .parent()
-                    .map(|p| p.to_string_lossy().into_owned())
-                    .filter(|s| !s.is_empty());
-                let name_color = status_color(f.status);
-                let path = f.path.clone();
-                div()
-                    .id(("push-file", i))
-                    .flex()
-                    .flex_row()
-                    .items_center()
-                    .gap_2()
-                    .mx(px(6.0))
-                    .px_2()
-                    .py_1()
-                    .rounded_md()
-                    .cursor_pointer()
-                    .hover(|s| s.bg(t.bg_mid))
-                    .child(div().flex_none().child(badge_inner(file_badge(&path), 2.0)))
-                    .child(div().flex_none().text_color(name_color).child(name))
-                    .when_some(dir, |d, dir| {
-                        d.child(
-                            div()
-                                .flex_1()
-                                .min_w_0()
-                                .truncate()
-                                .text_color(t.line_number)
-                                .child(SharedString::from(dir)),
-                        )
-                    })
-                    .on_mouse_down(
-                        MouseButton::Right,
-                        cx.listener(move |this, e: &gpui::MouseDownEvent, _w, cx| {
-                            this.open_menu(e.position, MenuTarget::PushFile(i), cx);
-                        }),
-                    )
-                    .into_any_element()
-            })
-            .collect();
-
-        let list = div()
-            .id("push-list")
-            .overflow_y_scroll()
-            .flex()
-            .flex_col()
-            .flex_1()
-            .py_1()
-            .child(
-                div()
-                    .px_3()
-                    .py_1()
-                    .text_color(t.line_number)
-                    .child(SharedString::from(if n == 0 {
-                        "Nothing to push.".to_string()
-                    } else {
-                        format!("{n} file{} to push", if n == 1 { "" } else { "s" })
-                    })),
-            )
-            .children(rows);
-
-        let footer = div()
-            .flex()
-            .flex_row()
-            .justify_end()
-            .gap_2()
-            .px_3()
-            .py_2()
-            .border_t_1()
-            .border_color(t.divider)
-            .child(
-                btn_secondary("push-cancel", "Cancel")
-                    .py_1p5()
-                    .on_mouse_down(
-                        MouseButton::Left,
-                        cx.listener(|this, _e, _w, cx| {
-                            this.close_modal_window(ModalKind::Push, cx);
-                        }),
-                    ),
-            )
-            .child(
-                btn_primary_state(
-                    "push",
-                    if self.pushing { "Pushing…" } else { "Push" },
-                    self.pushing,
-                )
-                .py_1p5()
-                .on_mouse_down(
-                    MouseButton::Left,
-                    cx.listener(|this, _e, _w, cx| this.do_push(cx)),
-                ),
-            );
-
-        let mut panel = div()
-            .size_full()
-            .flex()
-            .flex_col()
-            .bg(t.panel_bg)
-            .font_family(ui)
-            .text_size(px(theme::get().ui_font_size))
-            .text_color(t.text)
-            .child(list)
-            .child(footer);
-        // Right-click → View Diff menu renders inside THIS window (target = PushFile).
-        if matches!(
-            self.context_menu.as_ref().map(|m| &m.target),
-            Some(MenuTarget::PushFile(_))
-        ) {
-            panel = panel.child(self.render_context_menu(cx));
-        }
-        panel.into_any_element()
-    }
 
     /// "Create New Branch" dialog body (own native window). Name field (`branch_query`) +
     /// Checkout / Overwrite toggles + Cancel / Create.
